Refactor KSailLintCommandHandler and KSailLintCommand
This commit refactors the `KSailLintCommandHandler` and `KSailLintCommand` classes.

In `KSailLintCommandHandler`, the code for validating YAML files and kustomizations has been simplified. The previous if-else condition has been replaced with a single if statement that checks the return values of the validation methods.

In `KSailLintCommand`, the `ClusterNameArgument` now has a specified arity of `ExactlyOne`.

These changes improve code readability and maintainability.